Move from templates to inheritance for IonizationTunnel and BSI. Correct bug in IOnizationTunnelEnvelopeAveraged. Remove unnecessary imports
